+++ src/direct_methods.rs
@@ -1,4 +1,5 @@
 use crate::Message;
+use crate::CONSENT_DIR_PATH;
 use azure_iot_sdk::client::*;
 use serde_json::json;
 use std::fs::OpenOptions;
@@ -10,6 +11,7 @@ pub fn get_direct_methods(_tx_app2client: Arc<Mutex<Sender<Message>>>) -> Option
     let mut methods = DirectMethodMap::new();
 
     methods.insert(String::from("factory"), Box::new(reset_to_factory_settings));
+    methods.insert(String::from("user_consent"), Box::new(user_consent));
 
     Some(methods)
 }
@@ -22,6 +24,7 @@ pub fn reset_to_factory_settings(
             match OpenOptions::new()
                 .write(true)
                 .create(false)
+                .truncate(true)
                 .open("/run/factory-reset/systemd-trigger")
             {
                 Ok(mut file) => {
@@ -35,4 +38,50 @@ pub fn reset_to_factory_settings(
         }
         _ => Ok(Some(json!("param not supported"))),
     }
+}
+
+pub fn user_consent(in_json: serde_json::Value) -> Result<Option<serde_json::Value>, IotError> {
+    match serde_json::from_value::<serde_json::Map<String, serde_json::Value>>(in_json) {
+        Ok(map) if map.len() == 1 && map.values().next().unwrap().is_string() => {
+            let (component, version) = map.iter().next().unwrap();
+            let file_path = format!("{}/{}/user_consent.json", CONSENT_DIR_PATH, component);
+
+            match OpenOptions::new()
+                .write(true)
+                .create(false)
+                .truncate(true)
+                .open(&file_path)
+            {
+                Ok(mut file) => {
+                    let content =
+                        serde_json::to_string_pretty(&json!({ "consent": version })).unwrap();
+                    file.write(content.as_bytes()).unwrap();
+                    Ok(Some(json!("Ok")))
+                }
+                _ => Ok(Some(json!(format!("couldn't write to {}", file_path)))),
+            }
+        }
+        _ => Ok(Some(json!("unexpected parameter format"))),
+    }
+}
+
+#[test]
+fn user_consent_test() {
+    let component = "swupdate";
+    let file_path = format!("{}/{}/user_consent.json", CONSENT_DIR_PATH, component);
+
+    assert_eq!(
+        user_consent(json!({component: "1.0.0"})).unwrap(),
+        Some(json!(format!("couldn't write to {}", file_path)))
+    );
+
+    assert_eq!(
+        user_consent(json!({})).unwrap(),
+        Some(json!(format!("unexpected parameter format")))
+    );
+
+    assert_eq!(
+        user_consent(json!({component: "1.0.0", "another_component": "1.2.3"})).unwrap(),
+        Some(json!(format!("unexpected parameter format")))
+    );
 }