+++ src/main.rs
@@ -8,7 +8,6 @@ fn main() {
     } else {
         Builder::from_env(Env::default().default_filter_or("info")).init();
     }
-
     if let Err(e) = demo_portal_module::run() {
         error!("Application error: {}", e);