+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.0] Q2 2022
  - added user consent handling for "iot-hub-device-update"
  - bumped to latest azure-iot-sdk 0.5.5
+ - delete file content with OpenOptions "truncate"
 
 ## [0.2.0] Q1 2022
  - bumped to latest azure-iot-sdk 0.5.3